PSARC/2009/613 auditon(2) clarification 6884238 The audit_policy variable should be a uint32_t.

+++ usr/src/cmd/audit/audit.c
@@ -294,7 +294,7 @@ is_audit_control_ok(char *filename) {
 static boolean_t
 is_valid_zone(boolean_t show_err)
 {
-	long	policy;
+	uint32_t	policy;
 
 	if (auditon(A_GETPOLICY, (char *)&policy, 0) == -1) {
 		(void) fprintf(stderr, gettext(
